ktime: Get rid of the union and cleanup
ktime is a union because the initial implementation stored the time in
scalar nanoseconds on 64 bit machine and in a endianess optimized timespec
variant for 32bit machines. The Y2038 cleanup removed the timespec variant
and switched everything to scalar nanoseconds. The union remained, but
become completely pointless.

Get rid of the union and just keep ktime_t as simple typedef of type s64.

The conversion was done with coccinelle and some manual mopping up.

Change-Id: I22841d3f16aec9f6630c45e5e325822f1c6d73e1
Signed-off-by: Thomas Gleixner <tglx@linutronix.de>
Signed-off-by: Jason Edson <jaysonedson@gmail.com>