sched: Avoid placing task on isolated prev_cpu
When best energy efficient cpu is not found, then task gets
placed in its previous cpu, and it could be isolated.
For non-kernel threads, fallback mechanism selects unisolated cpu
which is extra overhead, and also unpinned kernel threads are gets
placed in the previous cpu till the energy efficient cpu found.
Which makes the isolated cpu to wakeup, though the kernel thread
can run on other cpus.

So make sure the fbt returns unisolated cpu, if the task's prev_cpu
is isolated.

Change-Id: Ia13cb08814141d1c08063c1aa9898f7a0101ecd5
Signed-off-by: Pavankumar Kondeti <pkondeti@codeaurora.org>
Signed-off-by: Lingutla Chandrasekhar <clingutla@codeaurora.org>
Signed-off-by: mydongistiny <jaysonedson@gmail.com>