sched: avoid energy evaluation for isolated prev_cpu
If task placement algorithm is not able to find efficient cpu,
it bias towards prev_cpu. If prev_cpu is isolated then, we use one
of non-isolated task allowed cpu and skip energy evaluations.
If we find efficient or backup cpu, prev_cpu is isolated then still it
is used for energy evaluations, which is not needed.

If prev_cpu isolated, set eenv flag independent of target_cpu to avoid
unnecessary energy evaluations.

Change-Id: I884d0e224b2f40157ef55fd429e138cf66eb6de1
Signed-off-by: Lingutla Chandrasekhar <clingutla@codeaurora.org>
Signed-off-by: Jason Edson <jaysonedson@gmail.com>