Makefile: LTO: Enable ThinLTO caching if path defined
Set KERNEL_THINLTO_CACHE path as the ThinLTO cache path if it's
defined. Speeds up linking quite a bit for parts that are the
same. Helps when building Android when no kernel changes have
been made.

Setup to work with either lld or gold as the linker.

Also limit the cache size to 5% of drive capacity or 5GB.

Change-Id: I333d08cf5d4e9a5b4c140c161d7d8773cacbd5fb
Signed-off-by: mydongistiny <jaysonedson@gmail.com>