iommu: msm: Rewrite to improve clarity and performance
This scope of this driver's lock usage is extremely wide, leading to
excessively long lock hold times. Other issues aside, this is the most
significant contributor to poor performance in this driver.

Fix everything by greatly reducing the scope of the locks used and by
using atomic reader/writer locks. The entire driver's code is refactored
and improved for better cleanliness and clarity.

Change-Id: Ifa966d457325cd88c6e302f9d6db1e4b36fb1179
Signed-off-by: Sultan Alsawaf <sultan@kerneltoast.com>
Signed-off-by: mydongistiny <jaysonedson@gmail.com>